Create README.md
This project is all about bike sales and purchases according to different age groups, region, gender, income and educational qualifications. The dataset used for this analysis link clearly defines the bike needs of different people based on the factors earlier listed. It was discovered that young people had the highest bike purchase as compared to others. this analysis was clearly illustrated and summarized in an excel dashboard which is attached to this project.